Continuous live

+++ hls-dl.py
@@ -1,23 +1,42 @@
 #!/usr/bin/env python3
 #
 #
+import time
 import os
 import m3u8
 import requests
 from argparse import ArgumentParser
+from hashlib import sha1
 
 
 CHUNK_SIZE = 10240
 
 
-def fetch_segment(url, output_dir):
-    """ Fetch a single MPEG-TS segment to `output_dir`.
-    """
+def fetch_key(url, output_dir):
+    local_path = os.path.join(output_dir, os.path.basename(url))
+    if os.path.exists(local_path):
+        return
+
     print("\t", os.path.basename(url))
     r = requests.get(url, stream=True)
     r.raise_for_status()
 
+    with open(local_path, 'wb') as f:
+        for chunk in r.iter_content(CHUNK_SIZE):
+            f.write(chunk)
+
+
+def fetch_segment(url, output_dir):
+    """ Fetch a single MPEG-TS segment to `output_dir`.
+    """
     local_path = os.path.join(output_dir, os.path.basename(url))
+    if os.path.exists(local_path):
+        return
+
+    print("\t", os.path.basename(url))
+    r = requests.get(url, stream=True)
+    r.raise_for_status()
+
     with open(local_path, 'wb') as f:
         for chunk in r.iter_content(CHUNK_SIZE):
             f.write(chunk)
@@ -27,11 +46,18 @@ def fetch_playlist(url, output_dir):
     """ Fetch an M3U8 playlist recursively and store the full
         contents in `output_dir`.
     """
-    print("Fetching", os.path.basename(url))
+    base = os.path.basename(url)
     r = requests.get(url)
     r.raise_for_status()
 
-    local_path = os.path.join(output_dir, os.path.basename(url))
+    base_sha = sha1(r.content).hexdigest()
+    local_path = os.path.join(output_dir, f"{base_sha}-{base}")
+
+    if base != "master.m3u8" and os.path.exists(local_path):
+        return
+
+    print("Fetching", base)
+
     with open(local_path, 'wb') as f:
         f.write(r.content)
 
@@ -40,6 +66,9 @@ def fetch_playlist(url, output_dir):
         for sp in p.playlists:
             fetch_playlist(sp.absolute_uri, output_dir)
     else:
+        for k in filter(None, p.keys):
+            fetch_key(k.absolute_uri, output_dir)
+
         for s in p.segments:
             fetch_segment(s.absolute_uri, output_dir)
 
@@ -60,4 +89,6 @@ def main():
 
 
 if __name__ == '__main__':
-    main()
+    while True:
+        main()
+        time.sleep(1)
